fix: handle fk constraints in equipment modules

+++ src/application/services/ProjectCategoryServices.ts
@@ -5,6 +5,7 @@ import IProjectCategoryRepository from "@/domain/repositories/iProjectCategoryRe
 import { ApplicationExceptionName } from "@/application/constants/applicationExceptionName";
 import ApplicationException from "@/application/exceptions/applicationException";
 import ProjectCategory from "@/domain/models/projectCategory";
+import IProjectRepository from "@/domain/repositories/iProjectRepository";
 import { IUnitOfWork } from "../unitOfWork/iUnitOfWork";
 
 @injectable()
@@ -12,6 +13,8 @@ class ProjectCategoryServices implements IProjectCategoryServices {
     constructor(
         @inject("ProjectCategoryRepository")
         private readonly projectCategoryRepository: IProjectCategoryRepository,
+        @inject("ProjectRepository")
+        private readonly projectRepository: IProjectRepository,
         @inject("SequelizeUnitOfWork")
         private readonly unitOfWork: IUnitOfWork
     ) { }
@@ -26,18 +29,25 @@ class ProjectCategoryServices implements IProjectCategoryServices {
     }
     async CreateAsync(data: ProjectCategoryDTO) {
         return await this.unitOfWork.execute(async (trx) => {
-            return (await this.projectCategoryRepository.Create(data, trx)) as ProjectCategoryDTO
+            return (await this.projectCategoryRepository.Create(data as ProjectCategory, trx)) as ProjectCategoryDTO
         })
     }
     async UpdateAsync(id: number, data: ProjectCategoryDTO) {
         return await this.unitOfWork.execute(async (trx) => {
-            if (!(await this.projectCategoryRepository.FindById(id) as ProjectCategory))
+            if (!(await this.projectCategoryRepository.FindById(id, trx) as ProjectCategory))
                 throw new ApplicationException(ApplicationExceptionName.NOT_FOUND, "No project category was found with the provided id", 404)
-            return (await this.projectCategoryRepository.Update(id, data, trx)) as ProjectCategoryDTO
+            return (await this.projectCategoryRepository.Update(id, data as ProjectCategory, trx)) as ProjectCategoryDTO
         })
     }
     async DeleteAsync(id: number) {
         return await this.unitOfWork.execute(async (trx) => {
+            if (!(await this.projectCategoryRepository.FindById(id, trx) as ProjectCategory))
+                throw new ApplicationException(ApplicationExceptionName.NOT_FOUND, "No project category was found with the provided id", 404)
+
+            const projectCount = await this.projectRepository.CountByProjectCategoryId(id, trx)
+            if (projectCount > 0)
+                throw new ApplicationException(ApplicationExceptionName.INVALID_OPERATION, "Project category is linked to projects and cannot be deleted", 400)
+
             return (await this.projectCategoryRepository.Delete(id, trx))
         })
     }
